+++ packages/polyfills/src/fetch.ts
@@ -1,1 +1,1 @@
-import 'whatwg-fetch';
+require('whatwg-fetch');